// src/app/registro/models/usuario.model.ts
/**
 * Modelado que contiene las propiedas para registrar a nu nuevo usuario
 */
export class UsuarioModel {
    nombre: string;
    apellido: string;
    genero: string;
    email: string;
    password: string;
    telefono: string;
    direccion: string;
    cp: string;
    sueldoMensual: number;
    idCarniceria: number;
    idEstado: number;
    idMunicipio: number;
    nombreEmpresa: string;
    role: string;
}